Fix incorrect importing
