+++ frontend/src/App.js
@@ -8,12 +8,29 @@ import Ranking from './components/Ranking';
 
 // Manage State
 import { useUserData } from './hooks/useUserData';
-import { setPlayArea, openQR, closeModal } from './helpers/dispatch';
 import { UserContext } from './contexts/userContext';
+import { ACTIONS } from './constants/ACTIONS';
 
 function App() {
   // Custom hook manages all global state through action suite, and fetches data relevant to user on initial pageload
   const { state, dispatch } = useUserData();
+
+  const setPlayArea = (playAreaId) => {
+    // Clear current play area in case of falsey input from toggle
+    if (!playAreaId) {
+      return dispatch({ type: ACTIONS.SET_CURRENT_PLAY_AREA, data: '' })
+    }
+    // Otherwise, set play area indo
+    return dispatch({ type: ACTIONS.SET_CURRENT_PLAY_AREA, data: playAreaId})
+  };
+  
+  const openQR = () => {
+  
+  };
+  
+  const closeModal = () => {
+    return dispatch({ type: ACTIONS.CLOSE_MODAL })
+  };
 
   return (
     <Router>
