+++ frontend/src/hooks/useUserData.js
@@ -36,4 +36,6 @@ const useUserData = () => {
 
 
 
-export { ACTIONS, initialState, useUserData}
+
+
+export { initialState, useUserData }
